requirements.txt
Django==1.8.3
Pygments==2.0.2
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
httpie==0.9.2
psycopg2==2.6.1
requests==2.7.0
static3==0.6.1
wsgiref==0.1.2
